server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.6
django-allauth==0.52.0
django-cors-headers==4.0.0
django-extensions==3.2.0
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
Faker==18.9.0
fonttools==4.39.4
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.1
oauthlib==3.2.2
openai==0.27.6
packaging==23.0
pandas==2.0.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
toposort==1.10
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
